create class for push notif
